0206/study01/pyproject.toml
[project]
name = "study01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.2",
    "kafka-python>=2.3.0",
    "pydantic-settings>=2.12.0",
]
